Move openj9 based impl of ExecuteCRIU into checkpoint
Moving ExecuteCRIU_OpenJ9 into io.openliberty.checkpoint
to simplify the code.  No longer need to allow alternative
implementations and it is easier to avoid a service lookup
for ExecuteCRIU implementations.
