+++ dev/com.ibm.websphere.appserver.features/visibility/public/checkpoint-1.0/io.openliberty.checkpoint-1.0.feature
@@ -6,8 +6,7 @@ IBM-ShortName: checkpoint-1.0
 IBM-Process-Types: server
 IBM-API-Package: io.openliberty.checkpoint.internal.criu; type="internal"
 Subsystem-Name: Checkpoint and Restore Support 1.0
--bundles=io.openliberty.checkpoint,\
- io.openliberty.checkpoint.openj9
+-bundles=io.openliberty.checkpoint
 kind=noship
 edition=full
 WLP-Activation-Type: parallel
