--- dev/io.openliberty.checkpoint.openj9/.gitignore
@@ -1,3 +1,0 @@
-/bin/
-/build/
-/generated/
