--- dev/io.openliberty.checkpoint.openj9/.settings/bndtools.core.prefs
@@ -1,2 +1,0 @@
-compileErrorAction=build
-eclipse.preferences.version=1
