--- dev/io.openliberty.checkpoint.openj9/bnd.bnd
@@ -1,32 +1,0 @@
-#*******************************************************************************
-# Copyright (c) 2020 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License v1.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-v10.html
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-#*******************************************************************************
--include= ~../cnf/resources/bnd/bundle.props
-
-bVersion: 1.0
-
-Bundle-Name: Open Liberty Checkpoint Support for Open J9
-Bundle-Description: Open Liberty Checkoint Support for Open J9, version ${bVersion}
-Bundle-SymbolicName: io.openliberty.checkpoint.openj9
-Private-Package: io.openliberty.checkpoint.openj9.internal
-Import-Package: org.eclipse.openj9.criu; resolution:=optional, *
-Bundle-Activator: io.openliberty.checkpoint.openj9.internal.J9CRIUSupportActivator
-
--buildpath: \
-	io.openliberty.checkpoint;version=latest,\
-	io.openliberty.org.eclipse.openj9.criu,\
-	com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
-	com.ibm.ws.logging.core, \
-	com.ibm.wsspi.org.osgi.service.component;version=latest,\
-	com.ibm.wsspi.org.osgi.service.component.annotations;version=latest,\
-	com.ibm.websphere.org.osgi.core
-
-
-
