+++ dev/io.openliberty.checkpoint/bnd.bnd
@@ -15,23 +15,24 @@ bVersion: 1.0
 Bundle-Name: Open Liberty Checkpoint Support
 Bundle-Description: Open Liberty Checkoint Support, version ${bVersion}
 Bundle-SymbolicName: io.openliberty.checkpoint
-Private-Package: io.openliberty.checkpoint.internal
+Private-Package: io.openliberty.checkpoint.internal.*
 Export-Package: io.openliberty.checkpoint.internal.criu
-Import-Package: org.apache.felix.service.command; resolution:=optional, *
+Import-Package: org.apache.felix.service.command; resolution:=optional, org.eclipse.openj9.criu; resolution:=optional, *
 
 -buildpath: \
-    com.ibm.websphere.org.osgi.core, \
-    com.ibm.ws.kernel.boot.core;version=latest, \
-    com.ibm.ws.org.apache.felix.gogo.runtime;version=latest, \
-    com.ibm.ws.org.osgi.annotation.versioning;version=latest, \
-    com.ibm.wsspi.org.osgi.service.component;version=latest, \
-    com.ibm.wsspi.org.osgi.service.component.annotations;version=latest, \
-    com.ibm.ws.kernel.feature.core;version=latest, \
-    com.ibm.ws.runtime.update;version=latest, \
-    com.ibm.ws.threading;version=latest, \
-    com.ibm.ws.logging.core,\
+	com.ibm.websphere.org.osgi.core,\
+	com.ibm.ws.kernel.boot.core;version=latest,\
+	com.ibm.ws.org.apache.felix.gogo.runtime;version=latest,\
+	com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
+	com.ibm.wsspi.org.osgi.service.component;version=latest,\
+	com.ibm.wsspi.org.osgi.service.component.annotations;version=latest,\
+	com.ibm.ws.kernel.feature.core;version=latest,\
+	com.ibm.ws.runtime.update;version=latest,\
+	com.ibm.ws.threading;version=latest,\
+	com.ibm.ws.logging.core,\
 	com.ibm.ws.kernel.service;version=latest,\
-	com.ibm.ws.org.objectweb.asm;version=latest
+	com.ibm.ws.org.objectweb.asm;version=latest,\
+	io.openliberty.org.eclipse.openj9.criu;version=latest
 
 -testpath: 	../build.sharedResources/lib/junit/old/junit.jar;version=file, \
 	com.ibm.ws.junit.extensions;version=latest, \
