+++ dev/io.openliberty.checkpoint/src/io/openliberty/checkpoint/internal/openj9/J9CRIUSupport.java
@@ -8,14 +8,11 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package io.openliberty.checkpoint.openj9.internal;
+package io.openliberty.checkpoint.internal.openj9;
 
 import java.io.File;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-
-import org.osgi.framework.BundleActivator;
-import org.osgi.framework.BundleContext;
 
 import com.ibm.ws.ffdc.annotation.FFDCIgnore;
 
@@ -23,53 +20,39 @@ import io.openliberty.checkpoint.internal.criu.CheckpointFailedException;
 import io.openliberty.checkpoint.internal.criu.CheckpointFailedException.Type;
 import io.openliberty.checkpoint.internal.criu.ExecuteCRIU;
 
-public class J9CRIUSupportActivator implements BundleActivator {
+public class J9CRIUSupport {
 
-    private BundleContext bundleContext;
-
-    @Override
     @FFDCIgnore({ ClassNotFoundException.class, InternalError.class })
-    public void start(BundleContext bc) {
-        bundleContext = bc;
+    public static ExecuteCRIU create() {
         try {
             Class<?> criuSupport = Class.forName("org.eclipse.openj9.criu.CRIUSupport");
             try {
                 Method supported = criuSupport.getDeclaredMethod("isCRIUSupportEnabled");
                 if (!(Boolean) supported.invoke(criuSupport)) {
-                    registerCRIUNotSupportedService(Type.UNSUPPORTED_DISABLED_IN_JVM,
-                                                    "Error: Must set the jvm option: -XX:+EnableCRIUSupport", null, 0);
-                    return;
+                    return createCRIUNotSupported(Type.UNSUPPORTED_DISABLED_IN_JVM,
+                                                  "Error: Must set the jvm option: -XX:+EnableCRIUSupport", null, 0);
                 }
             } catch (NoSuchMethodException | IllegalAccessException | InvocationTargetException hopefullyVeryRareException) {
-                registerCRIUNotSupportedService(Type.UNSUPPORTED,
-                                                "Error: Exception while invoking OpenJ9 method " +
-                                                                  "org.eclipse.openj9.criu.CRIUSupport.isCRIUSupportEnabled ",
-                                                hopefullyVeryRareException, 0);
-                return;
+                return createCRIUNotSupported(Type.UNSUPPORTED,
+                                              "Error: Exception while invoking OpenJ9 method " +
+                                                                "org.eclipse.openj9.criu.CRIUSupport.isCRIUSupportEnabled ",
+                                              hopefullyVeryRareException, 0);
             }
-            //Register a fully functional service. Yay!
-            bc.registerService(ExecuteCRIU.class, new ExecuteCRIU_OpenJ9(), null);
+            // return a fully functional implementation. Yay!
+            return new ExecuteCRIU_OpenJ9();
         } catch (ClassNotFoundException e) {
-            registerCRIUNotSupportedService(Type.UNSUPPORTED_IN_JVM, "There is no CRIU support in this JVM.", null, 0);
-            return;
+            return createCRIUNotSupported(Type.UNSUPPORTED_IN_JVM, "There is no CRIU support in this JVM.", null, 0);
         } catch (java.lang.InternalError ie) {
             if (ie.getCause() instanceof UnsatisfiedLinkError) {
-                registerCRIUNotSupportedService(Type.UNSUPPORTED_CRIU_NOT_INSTALLED,
-                                                "Error: CRIU is not installed on the platform", ie, 0);
-            } else {
-                registerCRIUNotSupportedService(Type.UNSUPPORTED, "An internal error was encountered: " + ie, ie, 0);
+                return createCRIUNotSupported(Type.UNSUPPORTED_CRIU_NOT_INSTALLED,
+                                              "Error: CRIU is not installed on the platform", ie, 0);
             }
-            return;
+            return createCRIUNotSupported(Type.UNSUPPORTED, "An internal error was encountered: " + ie, ie, 0);
         }
     }
 
-    @Override
-    public void stop(BundleContext bc) {
-
-    }
-
     /**
-     * Called when CRIU support is not present. Registers a ExecuteCRIU service which throws a variation of
+     * Called when CRIU support is not present. Returns a ExecuteCRIU service which throws a variation of
      * CheckpointFailedException from all public methods. The exception can be examined to get more info on why CRIU
      * support is missing,
      *
@@ -78,12 +61,12 @@ public class J9CRIUSupportActivator implements BundleActivator {
      * @param cause
      * @param errorCode
      */
-    private void registerCRIUNotSupportedService(CheckpointFailedException.Type type,
-                                                 String msg,
-                                                 Throwable cause,
-                                                 int errorCode) {
+    private static ExecuteCRIU createCRIUNotSupported(CheckpointFailedException.Type type,
+                                                      String msg,
+                                                      Throwable cause,
+                                                      int errorCode) {
         final CheckpointFailedException criuSupportException = new CheckpointFailedException(type, msg, cause, errorCode);
-        bundleContext.registerService(ExecuteCRIU.class, new ExecuteCRIU() {
+        return new ExecuteCRIU() {
             @Override
             public void dump(File imageDir, String logFileName, File workDir) throws CheckpointFailedException {
                 throw criuSupportException;
@@ -94,7 +77,7 @@ public class J9CRIUSupportActivator implements BundleActivator {
                 throw criuSupportException;
             }
 
-        }, null);
+        };
     }
 
 }
